% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_processSources.R
\name{buildDataFrame}
\alias{buildDataFrame}
\title{Build data.frame}
\usage{
buildDataFrame(source_info)
}
\arguments{
\item{source_info}{Single source info form getSources()}
}
\description{
Takes a column mapping and url to external data and builds a new data.frame
containing only matched columns
}
