% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{date2dateAndTime}
\alias{date2dateAndTime}
\title{date2dateAndTime}
\usage{
date2dateAndTime(data, postdate = TRUE)
}
\arguments{
\item{data}{Data frame to process}

\item{postdate}{Treat the post_date field to the same treatment?}
}
\description{
Some sources provide datetime information in the date field, this function
separates them.
}
