% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz2khz.R
\name{hz2khz}
\alias{hz2khz}
\title{Convert Hz to kHz}
\usage{
hz2khz(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
Data frame of processed data
}
\description{
Append a source column to the start of a data frame
}
