% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingestR}
\alias{ingestR}
\title{Ingest sources}
\usage{
ingestR(db = NULL, verbose = FALSE)
}
\arguments{
\item{db}{Database connection}

\item{verbose}{If TRUE says more about what's going on.}
}
\description{
Ingest data sources for audioBlast!
}
