% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{preprocessCoL}
\alias{preprocessCoL}
\title{Preprocess COL data}
\usage{
preprocessCoL(data)
}
\arguments{
\item{data}{dataframe of the CoL taxonomy.}
}
\value{
A reduced data.frame
}
\description{
Preprocess the Catalogue of Life data to remove unwanted columns and
simplify.
}
