% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{sourceR}
\alias{sourceR}
\title{sourceR}
\usage{
sourceR(source, data)
}
\arguments{
\item{source}{source string to append}

\item{data}{dataframe}
}
\value{
Data frame of processed data
}
\description{
Append a source column to the start of a data frame
}
