% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{streamTable}
\alias{streamTable}
\title{Stream a table to a file}
\usage{
streamTable(dir, type, table)
}
\arguments{
\item{dir}{Directory to stream to, made if it isn't there.}

\item{type}{Type of table, e.g. recordings or details.}

\item{table}{Rows to append, with the columns of getHeaders(type).}
}
\value{
Path of the file written to.
}
\description{
Appends a table's rows to the file of its type, which is given its header
the first time it is written to, so that a harvest too big to hold can be
written a page at a time (see xenocantoR()).
}
