% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_upload.R
\name{upload}
\alias{upload}
\title{Upload to audioBLAST!}
\usage{
upload(db, table, keys, data)
}
\arguments{
\item{db}{A database connection}

\item{table}{Database table to upload to}

\item{keys}{If these columns match existing data no INSERT will happen but an
UPDATE will}

\item{data}{A table of data to upload}
}
\description{
Function to upload data to audioBLAST!
}
