% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadAbiotic}
\alias{uploadAbiotic}
\title{Upload Abiotic}
\usage{
uploadAbiotic(db, i)
}
\arguments{
\item{db}{database connector}

\item{i}{dataframe of taxa to upload.}
}
\description{
Replaces the database abiotic table with contents of a data frame
}
