% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadDeploymentLocations}
\alias{uploadDeploymentLocations}
\title{uploadDeploymentLocations}
\usage{
uploadDeploymentLocations(db, i)
}
\arguments{
\item{db}{database connector}

\item{i}{dataframe of taxa to upload.}
}
\description{
Replaces the database deployment_locations table with contents of a data frame
}
