% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadDeployments}
\alias{uploadDeployments}
\title{uploadDeployments}
\usage{
uploadDeployments(db, i)
}
\arguments{
\item{db}{database connector}

\item{i}{dataframe of taxa to upload.}
}
\description{
Replaces the database taxa table with contents of a data frame
}
