% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadDevices}
\alias{uploadDevices}
\title{uploadDevices}
\usage{
uploadDevices(db, i)
}
\arguments{
\item{db}{database connector}

\item{i}{dataframe of taxa to upload.}
}
\description{
Replaces the database devices table with contents of a data frame
}
