% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadReferences}
\alias{uploadReferences}
\title{Upload References}
\usage{
uploadReferences(db, table)
}
\arguments{
\item{db}{database connector}

\item{table}{dataframe of references to upload, e.g. from bibtexR().}
}
\description{
Adds references from a data frame to the database references table, updating
references already in it. Empty values are uploaded as NULL.
}
