% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadSensors}
\alias{uploadSensors}
\title{uploadSensors}
\usage{
uploadSensors(db, i)
}
\arguments{
\item{db}{database connector}

\item{i}{dataframe of taxa to upload.}
}
\description{
Replaces the database sensors table with contents of a data frame
}
