% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_processDataSources.R
\name{verifyDataSource}
\alias{verifyDataSource}
\title{Data source verifications}
\usage{
verifyDataSource(source_info)
}
\arguments{
\item{source_info}{Single source info from getSources()}
}
\description{
Verifications that should be run on all data sources
}
