% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_genericVerifications.R
\name{verifyIdInMapping}
\alias{verifyIdInMapping}
\title{Verify a mapping against a column list}
\usage{
verifyIdInMapping(mapping)
}
\arguments{
\item{mapping}{List of mappings from getSources()}
}
\description{
Verifies that target columns used in a mapping exist in the
audioBLAST! database.
}
