% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_genericVerifications.R
\name{verifyMapping}
\alias{verifyMapping}
\title{Verify a mapping against a column list}
\usage{
verifyMapping(mapping, columns)
}
\arguments{
\item{mapping}{List of mappings from getSources()}

\item{columns}{Vector of column names to check against}
}
\description{
Verifies that target columns used in a mapping exist in the
audioBLAST! database.
}
