% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_processRecordings.R
\name{verifyRecordingData}
\alias{verifyRecordingData}
\title{Verify recording data}
\usage{
verifyRecordingData(data)
}
\arguments{
\item{data}{The data frame to be verified}
}
\description{
Verify integrity of recording data
}
