% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_processRecordings.R
\name{verifyRecordingSource}
\alias{verifyRecordingSource}
\title{Verify recording source}
\usage{
verifyRecordingSource(source_info)
}
\arguments{
\item{source_info}{A single source from getSources()}
}
\description{
Verify integrity of a recording source
}
