% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{col2flat}
\alias{col2flat}
\title{Flatten the Catalogue of Life taxonomy}
\usage{
col2flat(input)
}
\arguments{
\item{input}{dataframe of the CoL taxonomy}
}
\description{
Flatten the parent-child relationships in the Catalogue of Life taxonomy
}
