% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomise.R
\name{taxonomiseR}
\alias{taxonomiseR}
\title{taxonomiseR}
\usage{
taxonomiseR(input)
}
\arguments{
\item{input}{dataframe of taxa to process.}
}
\value{
Data frame of processed data
}
\description{
Processes a taxonomy file in the format proived from BioAcoustica.
}
