% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadRecordings}
\alias{uploadRecordings}
\title{uploadRecordings}
\usage{
uploadRecordings(db, i)
}
\arguments{
\item{db}{database connector}

\item{i}{dataframe of recordings to upload.}
}
\description{
Replaces the database recordings table with contents of a data frame
}
