% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadTaxa}
\alias{uploadTaxa}
\title{Upload Taxa}
\usage{
uploadTaxa(db, table)
}
\arguments{
\item{db}{database connector}

\item{table}{data.frame of taxa to upload.}
}
\description{
Replaces the database taxa table with contents of a data frame
}
