% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaders.R
\name{uploadTraits}
\alias{uploadTraits}
\title{Upload Traits}
\usage{
uploadTraits(db, table)
}
\arguments{
\item{db}{database connector}

\item{table}{dataframe of traits to upload.}
}
\description{
Replaces the database traits table with contents of a data frame
}
